+++ lib/launch/launch.dart
@@ -304,7 +304,7 @@ class Launch implements Disposable {
 
   bool get supportsRestart => false;
 
-  Future restart() => new Future.error('unsupported');
+  Future restart({ bool fullRestart: false }) => new Future.error('unsupported');
 
   void launchTerminated(int code, {bool quiet: false}) {
     if (isTerminated) return;